requirements.txt
# AWS SDK for Python (used for S3 uploads)
boto3==1.26.0

# Core Flask
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0

# MongoDB
Flask-PyMongo==2.3.0
pymongo==4.6.1

# Auth & Security
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1

# Email
Flask-Mail==0.9.1

# Realtime WebSockets (NO eventlet)
Flask-SocketIO==5.3.6
python-socketio==5.11.2
python-engineio==4.9.0

# Cloud Media
cloudinary==1.40.0

# Firebase Admin SDK
firebase-admin==7.1.0

# AI
openai==1.3.0

# Production servers (async/threading)
gunicorn==21.2.0
uvicorn==0.29.0
